Moved method modify to EventDataAccessObject from CommonEventFactory and implemented it properly.
