+++ src/entity/event/CommonEventFactory.java
@@ -7,7 +7,7 @@ import javax.ws.rs.core.MediaType;
 
 public class CommonEventFactory implements EventFactory {
 
-    static private int next_id = 0; //TODO: Add some check to make sure IDs are not repeated
+    static private int next_id = 0; //TODO: Add some check to make sure IDs are not repeated, maybe a list of in-use ids
     @Override
     public Event create(String name, String start, String end, String currency, String summary, Boolean isPrivate) {
         Client client = ClientBuilder.newClient();
@@ -41,19 +41,4 @@ public class CommonEventFactory implements EventFactory {
         return new CommonEvent(id, name, start, end, currency, summary, isPrivate);
     }
 
-    //TODO move this to DataAccessObject
-    public Event modify(String id, String name, String start, String end, String currency, String summary, Boolean isPrivate) {
-        Client client = ClientBuilder.newClient();
-        Entity payload = Entity.json("{  \"event\": {    \"name\": {      \"html\": \"&#60;p&#62;Some text&#60;/p&#62;\"    },    \"description\": {      \"html\": \"&#60;p&#62;Some text&#60;/p&#62;\"    },    \"start\": {      \"timezone\": \"UTC\",      \"utc\": \"2018-05-12T02:00:00Z\"    },    \"end\": {      \"timezone\": \"UTC\",      \"utc\": \"2018-05-12T02:00:00Z\"    },    \"currency\": \"USD\",    \"online_event\": false,    \"organizer_id\": \"\",    \"listed\": false,    \"shareable\": false,    \"invite_only\": false,    \"show_remaining\": true,    \"password\": \"12345\",    \"capacity\": 100,    \"is_reserved_seating\": true,    \"is_series\": true,    \"show_pick_a_seat\": true,    \"show_seatmap_thumbnail\": true,    \"show_colors_in_seatmap_thumbnail\": true  }}");
-        Response response = client.target("https://www.eventbriteapi.com/v3/events/" + id + "/")
-                .request(MediaType.APPLICATION_JSON_TYPE)
-                .header("Authorization", "Bearer PERSONAL_OAUTH_TOKEN")
-                .post(payload);
-
-        System.out.println("status: " + response.getStatus());
-        System.out.println("headers: " + response.getHeaders());
-        System.out.println("body:" + response.readEntity(String.class));
-        // TODO: Replace this with the Event retrieved from the data access object
-        return new CommonEvent("0", name, start, end, currency, summary, isPrivate);
-    }
 }
